Move PdfJs to vendors package
