+++ src/attachment/AttachmentLoader.tsx
@@ -8,8 +8,8 @@
 
 import React from 'react';
 
-import PdfJs from '../PdfJs';
 import Spinner from '../components/Spinner';
+import PdfJs from '../vendors/PdfJs';
 import AttachmentList from './AttachmentList';
 import FileItem from './FileItem';
