+++ src/attachment/FileItem.ts
@@ -6,7 +6,7 @@
  * @copyright 2019-2020 Nguyen Huu Phuoc <me@phuoc.ng>
  */
 
-import PdfJs from '../PdfJs';
+import PdfJs from '../vendors/PdfJs';
 
 interface FileItem {
     data: PdfJs.FileData;
