+++ src/layers/AnnotationLayer.tsx
@@ -8,9 +8,9 @@
 
 import React from 'react';
 
-import PdfJs from '../PdfJs';
 import ThemeContent from '../theme/ThemeContext';
 import getDestination from '../utils/getDestination';
+import PdfJs from '../vendors/PdfJs';
 import { SpecialLevel } from '../zoom/zoomingLevel';
 import './annotationLayer.less';
 import WithScale from './WithScale';
