+++ src/layers/CanvasLayer.tsx
@@ -8,8 +8,8 @@
 
 import React from 'react';
 
-import PdfJs from '../PdfJs';
 import ThemeContent from '../theme/ThemeContext';
+import PdfJs from '../vendors/PdfJs';
 import './canvasLayer.less';
 import WithScale from './WithScale';
