+++ src/layouts/Toolbar.tsx
@@ -31,7 +31,6 @@ import ZoomOutIcon from '../icons/ZoomOutIcon';
 import { RenderToolbarSlot } from './ToolbarSlot';
 import LocalizationContext from '../localization/LocalizationContext';
 import OpenFileButton from '../open/OpenFileButton';
-import PdfJs from '../PdfJs';
 import Modal from '../portal/Modal';
 import Position from '../portal/Position';
 import Tooltip from '../portal/Tooltip';
@@ -41,6 +40,7 @@ import SearchPopover from '../search/SearchPopover';
 import ScrollMode from '../ScrollMode';
 import SelectionMode from '../SelectionMode';
 import ThemeContent from '../theme/ThemeContext';
+import PdfJs from '../vendors/PdfJs';
 import { decrease, increase, SpecialLevel } from '../zoom/zoomingLevel';
 import ZoomPopover from '../zoom/ZoomPopover';
 import MoreActionsPopover from './MoreActionsPopover';
