+++ src/property/PropertiesLoader.tsx
@@ -8,9 +8,9 @@
 
 import React from 'react';
 
-import PdfJs from '../PdfJs';
 import Spinner from '../components/Spinner';
 import ThemeContent from '../theme/ThemeContext';
+import PdfJs from '../vendors/PdfJs';
 import PropertiesData from './PropertiesData';
 import './propertiesLoader.less';
