+++ src/property/PropertiesModal.tsx
@@ -11,11 +11,11 @@ import React from 'react';
 import PrimaryButton from '../components/PrimaryButton';
 import Separator from '../components/Separator';
 import LocalizationContext from '../localization/LocalizationContext';
-import PdfJs from '../PdfJs';
 import ThemeContent from '../theme/ThemeContext';
 import convertDate from '../utils/convertDate';
 import getFileName from '../utils/fileName';
 import getFileSize from '../utils/fileSize';
+import PdfJs from '../vendors/PdfJs';
 import PropertiesData from './PropertiesData';
 import PropertiesLoader from './PropertiesLoader';
 import './propertiesModal.less';
