+++ src/vendors/PdfJs.d.ts
